# api/openapi.yaml
openapi: 3.0.3
info:
  title: happyDeliver API
  description: Email Deliverability Testing Platform API
  version: 0.1.0
  contact:
    name: happyDomain team
    url: https://github.com/happyDomain/happydeliver
    email: contact+api@happydomain.org
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://spdx.org/licenses/AGPL-3.0-or-later.html
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
  /test:
    post:
      tags:
        - tests
      summary: Create a new deliverability test
      description: Generates a unique test email address for sending test emails. No database record is created until an email is received.
      operationId: createTest
      responses:
        '201':
          description: Test email address generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /test/{id}:
    get:
      tags:
        - tests
      summary: Get test status
      description: Check if a report exists for the given test ID (base32-encoded). Returns pending if no report exists, analyzed if a report is available.
      operationId: getTest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
            description: Base32-encoded test ID (with hyphens)
      responses:
        '200':
          description: Test status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /domain:
    post:
      tags:
        - tests
      summary: Test a domain's email configuration
      description: Analyzes DNS records (MX, SPF, DMARC, BIMI) for a domain without requiring an actual email to be sent. Returns results immediately.
      operationId: testDomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainTestRequest'
      responses:
        '200':
          description: Domain test completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainTestResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /blacklist:
    post:
      tags:
        - tests
      summary: Check an IP address against DNS blacklists
      description: Tests a single IP address (IPv4 or IPv6) against configured DNS-based blacklists (RBLs) without requiring an actual email to be sent. Returns results immediately.
      operationId: checkBlacklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlacklistCheckRequest'
      responses:
        '200':
          description: Blacklist check completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlacklistCheckResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Test:
      type: object
      required:
        - id
        - email
        - status
      properties:
        id:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Unique test identifier (base32-encoded with hyphens)
          example: "krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a"
        email:
          type: string
          format: email
          description: Unique test email address
          example: "test-krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a@example.com"
        status:
          type: string
          enum: [pending, analyzed]
          description: Current test status (pending = no report yet, analyzed = report available)
          example: "analyzed"
    TestResponse:
      type: object
      required:
        - id
        - email
        - status
      properties:
        id:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Unique test identifier (base32-encoded with hyphens)
          example: "krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a"
        email:
          type: string
          format: email
          example: "test-krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a@example.com"
        status:
          type: string
          enum: [pending]
          example: "pending"
        message:
          type: string
          example: "Send your test email to the address above"
...
    DNSResults:
      type: object
      required:
        - from_domain
      properties:
        from_domain:
          type: string
          description: From Domain name
          example: "example.com"
        rp_domain:
          type: string
          description: Return Path Domain name
          example: "example.com"
        from_mx_records:
          type: array
          items:
            $ref: '#/components/schemas/MXRecord'
          description: MX records for the From domain
        rp_mx_records:
          type: array
          items:
            $ref: '#/components/schemas/MXRecord'
          description: MX records for the Return-Path domain
        spf_records:
          type: array
          items:
            $ref: '#/components/schemas/SPFRecord'
          description: SPF records found (includes resolved include directives)
        dkim_records:
          type: array
          items:
            $ref: '#/components/schemas/DKIMRecord'
          description: DKIM records found
        dmarc_record:
          $ref: '#/components/schemas/DMARCRecord'
        bimi_record:
          $ref: '#/components/schemas/BIMIRecord'
        ptr_records:
          type: array
          items:
            type: string
          description: PTR (reverse DNS) records for the sender IP address
          example: ["mail.example.com", "smtp.example.com"]
        ptr_forward_records:
          type: array
          items:
            type: string
          description: A or AAAA records resolved from the PTR hostnames (forward confirmation)
          example: ["192.0.2.1", "2001:db8::1"]
        errors:
          type: array
          items:
            type: string
          description: DNS lookup errors
    MXRecord:
      type: object
      required:
        - host
        - priority
        - valid
      properties:
        host:
          type: string
          description: MX hostname
          example: "mail.example.com"
        priority:
          type: integer
          format: uint16
          description: MX priority (lower is higher priority)
          example: 10
        valid:
          type: boolean
          description: Whether the MX record is valid
          example: true
        error:
          type: string
          description: Error message if validation failed
          example: "Failed to lookup MX records"
    SPFRecord:
      type: object
      required:
        - valid
      properties:
        domain:
          type: string
          description: Domain this SPF record belongs to
          example: "example.com"
        record:
          type: string
          description: SPF record content
          example: "v=spf1 include:_spf.example.com ~all"
        valid:
          type: boolean
          description: Whether the SPF record is valid
          example: true
        all_qualifier:
          type: string
          enum: ["+", "-", "~", "?"]
          description: "Qualifier for the 'all' mechanism: + (pass), - (fail), ~ (softfail), ? (neutral)"
          example: "~"
        error:
          type: string
          description: Error message if validation failed
          example: "No SPF record found"
    DKIMRecord:
      type: object
      required:
        - selector
        - domain
        - valid
      properties:
        selector:
          type: string
          description: DKIM selector
          example: "default"
        domain:
          type: string
          description: Domain name
          example: "example.com"
        record:
          type: string
          description: DKIM record content
          example: "v=DKIM1; k=rsa; p=MIGfMA0GCSqGSIb3DQEBAQUAA..."
        valid:
          type: boolean
          description: Whether the DKIM record is valid
          example: true
        error:
          type: string
          description: Error message if validation failed
          example: "No DKIM record found"
    DMARCRecord:
      type: object
      required:
        - valid
      properties:
        record:
          type: string
          description: DMARC record content
          example: "v=DMARC1; p=quarantine; rua=mailto:dmarc@example.com"
        policy:
          type: string
          enum: [none, quarantine, reject, unknown]
          description: DMARC policy
          example: "quarantine"
        subdomain_policy:
          type: string
          enum: [none, quarantine, reject, unknown]
          description: DMARC subdomain policy (sp tag) - policy for subdomains if different from main policy
          example: "quarantine"
        percentage:
          type: integer
          minimum: 0
          maximum: 100
          description: Percentage of messages subjected to filtering (pct tag, default 100)
          example: 100
        spf_alignment:
          type: string
          enum: [relaxed, strict]
          description: SPF alignment mode (aspf tag)
          example: "relaxed"
        dkim_alignment:
          type: string
          enum: [relaxed, strict]
          description: DKIM alignment mode (adkim tag)
          example: "relaxed"
        valid:
          type: boolean
          description: Whether the DMARC record is valid
          example: true
        error:
          type: string
          description: Error message if validation failed
          example: "No DMARC record found"
    BIMIRecord:
      type: object
      required:
        - selector
        - domain
        - valid
      properties:
        selector:
          type: string
          description: BIMI selector
          example: "default"
        domain:
          type: string
          description: Domain name
          example: "example.com"
        record:
          type: string
          description: BIMI record content
          example: "v=BIMI1; l=https://example.com/logo.svg"
        logo_url:
          type: string
          format: uri
          description: URL to the brand logo (SVG)
          example: "https://example.com/logo.svg"
        vmc_url:
          type: string
          format: uri
          description: URL to Verified Mark Certificate (optional)
          example: "https://example.com/vmc.pem"
        valid:
          type: boolean
          description: Whether the BIMI record is valid
          example: true
        error:
          type: string
          description: Error message if validation failed
          example: "No BIMI record found"
    BlacklistCheck:
      type: object
      required:
        - rbl
        - listed
      properties:
        rbl:
          type: string
          description: RBL/DNSBL name
          example: "zen.spamhaus.org"
        listed:
          type: boolean
          description: Whether IP is listed
          example: false
        response:
          type: string
          description: RBL response code or message
          example: "127.0.0.2"
        error:
          type: string
          description: RBL error if any
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "not_found"
        message:
          type: string
          description: Human-readable error message
          example: "Test not found"
        details:
          type: string
          description: Additional error details
    DomainTestRequest:
      type: object
      required:
        - domain
      properties:
        domain:
          type: string
          pattern: '^[a-zA-Z0-9]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(\.[a-zA-Z0-9]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$'
          description: Domain name to test (e.g., example.com)
          example: "example.com"
    DomainTestResponse:
      type: object
      required:
        - domain
        - score
        - grade
        - dns_results
      properties:
        domain:
          type: string
          description: The tested domain name
          example: "example.com"
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall domain configuration score (0-100)
          example: 85
        grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score
          example: "A"
        dns_results:
          $ref: '#/components/schemas/DNSResults'
    BlacklistCheckRequest:
      type: object
      required:
        - ip
      properties:
        ip:
          type: string
          description: IPv4 or IPv6 address to check against blacklists
          example: "192.0.2.1"
          pattern: '^([0-9]{1,3}\.){3}[0-9]{1,3}$|^([0-9a-fA-F]{0,4}:){7}[0-9a-fA-F]{0,4}$|^::([0-9a-fA-F]{0,4}:){0,6}[0-9a-fA-F]{0,4}$|^([0-9a-fA-F]{0,4}:){1,6}:([0-9a-fA-F]{0,4}:){0,5}[0-9a-fA-F]{0,4}$'
    BlacklistCheckResponse:
      type: object
      required:
        - ip
        - blacklists
        - listed_count
        - score
        - grade
      properties:
        ip:
          type: string
          description: The IP address that was checked
          example: "192.0.2.1"
        blacklists:
          type: array
          items:
            $ref: '#/components/schemas/BlacklistCheck'
          description: List of blacklist check results
        listed_count:
          type: integer
          description: Number of blacklists that have this IP listed
          example: 0
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Blacklist score (0-100, higher is better)
          example: 100
        grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score
          example: "A+"
        whitelists:
          type: array
          items:
            $ref: '#/components/schemas/BlacklistCheck'
          description: List of DNS whitelist check results (informational only)
...